+++ tests/Helpers/Plugins/RenderPluginsTest.php
@@ -3,8 +3,10 @@
 namespace Larrock\Core\Tests\Helpers\Plugins;
 
 use Larrock\ComponentBlocks\LarrockComponentBlocksServiceProvider;
+use Larrock\ComponentBlocks\Models\Blocks;
 use Larrock\Core\LarrockCoreServiceProvider;
 use Larrock\Core\Tests\DatabaseTest\CreateBlocksDatabase;
+use Larrock\Core\Tests\DatabaseTest\CreateMediaDatabase;
 use Orchestra\Testbench\TestCase;
 use Larrock\Core\Helpers\Plugins\RenderPlugins;
 use Illuminate\Support\Facades\DB;
@@ -19,6 +21,12 @@ class RenderPluginsTest extends TestCase
         parent::setUp();
 
         $this->RenderPlugins = new RenderPlugins('html', 'model');
+
+        $seed = new CreateBlocksDatabase();
+        $seed->setUpBlocksDatabase();
+
+        $seed = new CreateMediaDatabase();
+        $seed->setUpMediaDatabase();
     }
 
     public function tearDown()
@@ -55,9 +63,6 @@ class RenderPluginsTest extends TestCase
 
     public function testRenderBlocks()
     {
-        $seed = new CreateBlocksDatabase();
-        $seed->setUpBlocksDatabase();
-
         DB::connection()->table('blocks')->insert([
             'title' => 'test',
             'description' => 'test',
@@ -66,6 +71,46 @@ class RenderPluginsTest extends TestCase
 
         $this->RenderPlugins = new RenderPlugins('{Блок[default]=render}', 'model');
         $test = $this->RenderPlugins->renderBlocks();
+        $this->assertNotEmpty($test->rendered_html);
+    }
+
+    public function testRenderImageGallery()
+    {
+        $model = new Blocks();
+        $this->RenderPlugins = new RenderPlugins('{Фото[news]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderImageGallery();
+        $this->assertNotEmpty($test->rendered_html);
+
+        $this->RenderPlugins = new RenderPlugins('{Фото[nonFancy]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderImageGallery();
+        $this->assertNotEmpty($test->rendered_html);
+
+        $this->RenderPlugins = new RenderPlugins('{Фото[newsDescription]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderImageGallery();
+        $this->assertNotEmpty($test->rendered_html);
+
+        $this->RenderPlugins = new RenderPlugins('{Фото[blocks]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderImageGallery();
+        $this->assertNotEmpty($test->rendered_html);
+
+        $this->RenderPlugins = new RenderPlugins('{Фото[blocksBig]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderImageGallery();
+        $this->assertNotEmpty($test->rendered_html);
+
+        $this->RenderPlugins = new RenderPlugins('{Фото[sert]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderImageGallery();
+        $this->assertNotEmpty($test->rendered_html);
+    }
+
+    public function testRenderFilesGallery()
+    {
+        $model = new Blocks();
+        $this->RenderPlugins = new RenderPlugins('{Файлы[default]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderFilesGallery();
+        $this->assertNotEmpty($test->rendered_html);
+
+        $this->RenderPlugins = new RenderPlugins('{Файлы[price]=render}', $model->find(1));
+        $test = $this->RenderPlugins->renderFilesGallery();
         $this->assertNotEmpty($test->rendered_html);
     }
 }